init ninja extra, asgi
