+++ .gitignore
@@ -205,3 +205,5 @@ pyrightconfig.json
 .ionide
 
 # End of https://www.toptal.com/developers/gitignore/api/python,django,visualstudiocode
+
+.DS_Store
