+++ pyproject.toml
@@ -7,11 +7,13 @@ readme = "README.md"
 
 [tool.poetry.dependencies]
 python = "^3.13"
-django = "^5.1.5"
-djangorestframework = "^3.15.2"
-py-eureka-client = "^0.11.13"
+django = "^5.1.6"
+django-ninja = "^1.3.0"
+django-ninja-extra = "^0.22.3"
 psycopg2-binary = "^2.9.10"
 pytest-django = "^4.9.0"
+py-eureka-client = "^0.11.13"
+channels = "^4.2.0"
 
 
 [build-system]
